fix(traverseschema.js): skip generating files for certain keywords in the schema
When traversing the schemas, some objects and arrays are misinterpreted as schemas to be documented
rather than as generic structures used to define the schemas. Namely: 'examples', 'required', and
'properties'. Including these structures in the loadedschemas list generates a number of extranious
files with little value to the generated documentation.

BREAKING CHANGE: The extranious files, if they have some other value, will not be generated and
there is no mechanism to continue to g