fix(lib/traverseschema.js): filter out skippable keywords based on parent's type
'properties', 'required' and 'examples' should be skipped only when they are part of the 'object' or
'array' schema type. These should not be ommited when then are part of 'properties' object, for
instance.

BREAKING CHANGE: If schema contained 'properties', 'required' or 'examples' as a part of JSON schema
(i.e. it was described as 'properties' within an object of 'object' type), these will be included
into output.

fix #282